+++ Cargo.toml
@@ -2,6 +2,8 @@
 name = "flichess"
 version = "0.1.0"
 edition = "2021"
+build = "build.rs"
+default-run = "flichess"
 
 [profile.release]
 lto = true
@@ -16,6 +18,9 @@ rustyline = "14.0.0"
 
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
+
+[build-dependencies]
+vergen-git2 = { version = "1.0.0", features = ["build", "cargo"] }
 
 [[bench]]
 name = "pawn_movegen"
