+++ src/lib.rs
@@ -1,10 +1,9 @@
+mod util;
+
 pub mod bitboard;
-pub mod board;
 pub mod chess;
 pub mod fen;
-pub mod magic;
-pub mod magic_finder;
-pub mod magic_gen;
 pub mod movegen;
 pub mod moves;
 pub mod position;
+pub mod uci;
