+++ src/movegen/magic_gen.rs
@@ -1,5 +1,4 @@
-use crate::bitboard::Bitboard;
-use crate::magic::Magic;
+use crate::{bitboard::Bitboard, movegen::Magic};
 
 #[rustfmt::skip]
 pub const BISHOP_MAGICS: [Magic; 64] = [
