+++ src/position.rs
@@ -2,7 +2,6 @@ use std::num::NonZeroU32;
 
 use crate::{
     bitboard::Bitboard,
-    board::Board,
     chess::{CastleRights, Color, File, Piece, Role, Square},
     movegen::{between, bishop_rays, get_knight_moves, get_pawn_attacks, rook_rays},
     moves::Move,
@@ -18,7 +17,13 @@ pub struct State {
 
 #[derive(Debug, Clone, Copy)]
 pub struct Position {
-    pub board: Board,
+    // pub board: Board,
+    pub by_color: [Bitboard; Color::NUM],
+    pub by_role: [Bitboard; Role::NUM],
+    pub occupancy: Bitboard,
+    pub checkers: Bitboard,
+    pub pinned: Bitboard,
+
     pub side: Color,
     pub castling: CastleRights,
     pub ep_square: Option<Square>,
@@ -30,7 +35,12 @@ pub struct Position {
 impl Position {
     pub fn new() -> Position {
         Position {
-            board: Board::new(),
+            // board: Board::new(),
+            by_color: [Bitboard::EMPTY; Color::NUM],
+            by_role: [Bitboard::EMPTY; Role::NUM],
+            occupancy: Bitboard::EMPTY,
+            checkers: Bitboard::EMPTY,
+            pinned: Bitboard::EMPTY,
             side: Color::White,
             castling: CastleRights::all(),
             ep_square: None,
@@ -49,33 +59,86 @@ impl Default for Position {
 
 impl Position {
     #[inline]
+    pub fn color_at(&self, sq: Square) -> Option<Color> {
+        self.by_color
+            .iter()
+            .position(|bb| bb.contains(sq))
+            .map(|idx| Color::new(idx as u8))
+    }
+
+    #[inline]
+    pub fn role_at(&self, sq: Square) -> Option<Role> {
+        if self.occupancy.contains(sq) {
+            self.by_role
+                .iter()
+                .position(|bb| bb.contains(sq))
+                .map(|idx| Role::new(idx as u8))
+        } else {
+            None
+        }
+    }
+
+    #[inline]
+    pub fn piece_at(&self, sq: Square) -> Option<Piece> {
+        self.role_at(sq).map(|role| Piece {
+            color: self.color_at(sq).unwrap(),
+            role,
+        })
+    }
+
+    #[inline]
+    pub fn by_color_role(&self, color: Color, role: Role) -> Bitboard {
+        self.by_color[color as usize] & self.by_role[role as usize]
+    }
+
+    #[inline]
+    pub fn king_of(&self, color: Color) -> Bitboard {
+        self.by_color_role(color, Role::King)
+    }
+
+    #[inline]
     pub fn us(&self) -> Bitboard {
-        self.board.by_color(self.side)
+        self.by_color[self.side as usize]
     }
 
     #[inline]
     pub fn them(&self) -> Bitboard {
-        self.board.by_color(self.side.opponent())
+        self.by_color[self.side.opponent() as usize]
     }
 
     #[inline]
     pub fn our(&self, role: Role) -> Bitboard {
-        self.board.by_color_role(self.side, role)
+        self.by_color_role(self.side, role)
     }
 
     #[inline]
     pub fn their(&self, role: Role) -> Bitboard {
-        self.board.by_color_role(self.side.opponent(), role)
+        self.by_color_role(self.side.opponent(), role)
     }
 
     #[inline]
     pub fn our_king(&self) -> Bitboard {
-        self.board.king_of(self.side)
+        self.by_color_role(self.side, Role::King)
     }
 
     #[inline]
     pub fn their_king(&self) -> Bitboard {
-        self.board.king_of(self.side.opponent())
+        self.by_color_role(self.side.opponent(), Role::King)
+    }
+
+    #[inline]
+    pub fn discard(&mut self, sq: Square) {
+        self.by_color.iter_mut().for_each(|bb| bb.clear(sq));
+        self.by_role.iter_mut().for_each(|bb| bb.clear(sq));
+        self.occupancy.clear(sq);
+    }
+
+    #[inline]
+    pub fn set(&mut self, sq: Square, Piece { color, role }: Piece) {
+        self.discard(sq);
+        self.by_color[color as usize].set(sq);
+        self.by_role[role as usize].set(sq);
+        self.occupancy.set(sq);
     }
 
     #[inline]
@@ -83,42 +146,42 @@ impl Position {
         let from = mv.from();
         let to = mv.to();
 
-        let piece = self.board.piece_at(from).unwrap();
+        let piece = self.piece_at(from).unwrap();
 
         // do move and capture
         let captured = if piece.role == Role::Pawn && self.ep_square == Some(to) {
             // unwrapping is safe here because we know ep_square is never at the edge of the board
             let captured_pawn_square = to.down(self.side).unwrap();
-            let captured = self.board.piece_at(captured_pawn_square);
-            self.board.discard(from);
-            self.board.discard(captured_pawn_square);
-            self.board.set(to, piece);
+            let captured = self.piece_at(captured_pawn_square);
+            self.discard(from);
+            self.discard(captured_pawn_square);
+            self.set(to, piece);
             captured
         } else {
-            let captured = self.board.piece_at(to);
-            self.board.discard(from);
-            self.board.set(to, piece);
+            let captured = self.piece_at(to);
+            self.discard(from);
+            self.set(to, piece);
             captured
         };
 
         if let Some(promotion) = mv.promotion() {
             let promoted = Piece::new(self.side, promotion);
-            self.board.set(to, promoted);
+            self.set(to, promoted);
         }
 
         // if a castling move, move the rook
         if piece.role == Role::King && from.file().direction(to.file()) == 2 {
             let rook_from = Square::make(File::H, self.side.back_rank());
             let rook_to = Square::make(File::F, self.side.back_rank());
-            let rook = self.board.piece_at(rook_from).unwrap();
-            self.board.discard(rook_from);
-            self.board.set(rook_to, rook);
+            let rook = self.piece_at(rook_from).unwrap();
+            self.discard(rook_from);
+            self.set(rook_to, rook);
         } else if piece.role == Role::King && from.file().direction(to.file()) == -2 {
             let rook_from = Square::make(File::A, self.side.back_rank());
             let rook_to = Square::make(File::D, self.side.back_rank());
-            let rook = self.board.piece_at(rook_from).unwrap();
-            self.board.discard(rook_from);
-            self.board.set(rook_to, rook);
+            let rook = self.piece_at(rook_from).unwrap();
+            self.discard(rook_from);
+            self.set(rook_to, rook);
         }
 
         // self.history.push(State {
@@ -186,17 +249,17 @@ impl Position {
         // we update side at the very end of make move,
         // so we're checking agaings the opponent king
 
-        *self.board.checkers_mut() = Bitboard::EMPTY;
-        *self.board.pinned_mut() = Bitboard::EMPTY;
+        self.checkers = Bitboard::EMPTY;
+        self.pinned = Bitboard::EMPTY;
 
         let dest_bb = Bitboard::from(mv.to());
 
         let ksq = Square::new_unchecked(self.their_king().0.trailing_zeros() as u8);
 
         if piece == Role::Knight {
-            *self.board.checkers_mut() |= get_knight_moves(ksq) & dest_bb;
+            self.checkers |= get_knight_moves(ksq) & dest_bb;
         } else if piece == Role::Pawn {
-            *self.board.checkers_mut() |= get_pawn_attacks(ksq, self.side.opponent()) & dest_bb;
+            self.checkers |= get_pawn_attacks(ksq, self.side.opponent()) & dest_bb;
         }
 
         let bishop_attackers = (self.our(Role::Bishop) | self.our(Role::Queen)) & bishop_rays(ksq);
@@ -204,28 +267,28 @@ impl Position {
         let attackers = bishop_attackers | rook_attackers;
 
         for sq in attackers {
-            let btw = between(ksq, sq) & self.board.occupancy();
+            let btw = between(ksq, sq) & self.occupancy;
 
             if btw == Bitboard::EMPTY {
-                *self.board.checkers_mut() |= Bitboard::from(sq);
+                self.checkers |= Bitboard::from(sq);
             } else if btw.count() == 1 {
                 let them = self.them();
-                *self.board.pinned_mut() |= btw & them
+                self.pinned |= btw & them
             }
         }
     }
 
     pub fn refresh_checks_and_pins(&mut self) {
         // fully refresh checks and pins for the current side
-        *self.board.checkers_mut() = Bitboard::EMPTY;
-        *self.board.pinned_mut() = Bitboard::EMPTY;
+        self.checkers = Bitboard::EMPTY;
+        self.pinned = Bitboard::EMPTY;
 
         let ksq = Square::new_unchecked(self.our_king().0.trailing_zeros() as u8);
 
         let knight_attackers = self.their(Role::Knight) & get_knight_moves(ksq);
         let pawn_attackers = self.their(Role::Pawn) & get_pawn_attacks(ksq, self.side.opponent());
 
-        *self.board.checkers_mut() |= knight_attackers | pawn_attackers;
+        self.checkers |= knight_attackers | pawn_attackers;
 
         let bishop_attackers =
             (self.their(Role::Bishop) | self.their(Role::Queen)) & bishop_rays(ksq);
@@ -233,12 +296,12 @@ impl Position {
 
         let attackers = bishop_attackers | rook_attackers;
         for sq in attackers {
-            let btw = between(ksq, sq) & self.board.occupancy();
+            let btw = between(ksq, sq) & self.occupancy;
             if btw == Bitboard::EMPTY {
-                *self.board.checkers_mut() |= Bitboard::from(sq);
+                self.checkers |= Bitboard::from(sq);
             } else if btw.count() == 1 {
                 let us = self.us();
-                *self.board.pinned_mut() |= btw & us;
+                self.pinned |= btw & us;
             }
         }
     }
